GoogleSTTClient/sttclient/requirements.txt
functions-framework==3.5.0
google-cloud-speech==2.7.0